pyproject.toml
[project]
name = "fossuok-qr"
version = "0.1.0"
description = "QR-based event registration system for FOSSUoK"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]==0.129.2",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "maileroo>=1.0.0",
    "pillow>=12.1.1",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "qrcode>=8.2",
    "sqlalchemy>=2.0.46",
]

[project.scripts]
app = "main:app"

[tool.vercel.scripts]
build = "python build.py"
